+++ src/main/kotlin/org/test/anymindtask/strategy/PricingStrategy.kt
@@ -1,7 +1,8 @@
 package org.test.anymindtask.strategy
 
 import org.test.anymindtask.model.Payment
+import java.math.BigDecimal
 
 interface PricingStrategy {
-    fun calculateFinalPriceAndPoints(payment: Payment): Pair<Double, Double>
+    fun calculateFinalPriceAndPoints(payment: Payment): Pair<BigDecimal, BigDecimal>
 }
